+++ MsaasBackend/Controllers/Admin/PhysiciansController.cs
@@ -5,6 +5,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
+using MsaasBackend.Helpers;
 using MsaasBackend.Models;
 
 namespace MsaasBackend.Controllers.Admin
