+++ MsaasBackend/Controllers/MedicalRecordsController.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
+using MsaasBackend.Helpers;
 
 namespace MsaasBackend.Controllers
 {
