+++ MsaasBackend/Controllers/Physicians/MedicalRecordsController.cs
@@ -6,6 +6,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
+using MsaasBackend.Helpers;
 using MsaasBackend.Models;
 
 namespace MsaasBackend.Controllers.Physicians
